auth_core/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.6.1
redis==5.0.1
python-multipart==0.0.20
bcrypt==4.0.1